Add ArrangeViewCanvas
